Split up vocab.py to avoid cyclic imports
